Render Article detail with unsafe HTML
