+++ Sources/App/configure.swift
@@ -7,14 +7,18 @@ import Vapor
 public func configure(_ app: Application) async throws {
   app.middleware.use(FileMiddleware(publicDirectory: app.directory.publicDirectory))
   
-  app.databases.use(DatabaseConfigurationFactory.postgres(configuration: .init(
-    hostname: AppEnvironment.hostname,
-    port: AppEnvironment.port,
-    username: AppEnvironment.username,
-    password: AppEnvironment.password,
-    database: AppEnvironment.database,
-    tls: .prefer(try .init(configuration: .clientDefault)))
-  ), as: .psql)
+  app.databases.use(
+    DatabaseConfigurationFactory.postgres(
+      configuration: .init(
+        hostname: AppEnvironment.hostname,
+        port: AppEnvironment.port,
+        username: AppEnvironment.username,
+        password: AppEnvironment.password,
+        database: AppEnvironment.database,
+        tls: .prefer(try .init(configuration: .clientDefault)))
+    ),
+    as: .psql
+  )
   
   app.migrations.add(CreateArticle())
   app.migrations.add(SeedArticles())
